fix(ios): satisfy SwiftLint on mobile feature files
Rename single-character initials bindings in MeProfile and extract item detail row builders so ItemDetailView stays within type_body_length limits.
